test_project/requirements.txt
Django==5.2.6
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.10
tortoise-orm[asyncpg]==0.25.1
factory_boy==3.3.3